mypage 주문내역 완성

+++ Gogo/src/gogo/order/controller/PayInsertController.java
@@ -1,6 +1,7 @@
 package gogo.order.controller;
 
 import java.io.IOException;
+import java.util.ArrayList;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -11,6 +12,7 @@ import javax.servlet.http.HttpServletResponse;
 import gogo.order.dao.BasketDao;
 import gogo.order.dao.BuyDao;
 import gogo.order.dao.PayDao;
+import gogo.order.vo.BasketVo;
 import gogo.order.vo.BuyVo;
 import gogo.order.vo.PayVo;
 
@@ -26,11 +28,14 @@ public class PayInsertController extends HttpServlet{
 
 		// 장바구니 비우기
 		BasketDao dao = BasketDao.getInstance();
-		int nn = dao.deleteAll(id);
-		if(nn <= 0) {
-			// 오류처리
-			ok = false;
-			System.out.println("장바구니 비우기 실패");
+		ArrayList<BasketVo> list = dao.getBasket(id);
+		if(list.size() > 0) {
+			int nn = dao.deleteAll(id);
+			if(nn <= 0) {
+				// 오류처리
+				ok = false;
+				System.out.println("장바구니 비우기 실패");
+			}
 		}
 		
 		// buy 테이블 배송주소 수정
@@ -46,7 +51,7 @@ public class PayInsertController extends HttpServlet{
 		// pay 테이블에 DB 추가
 		String pay_how = req.getParameter("payhow");
 		int pay_sum = Integer.parseInt(req.getParameter("pay_sum"));
-		PayVo pvo = new PayVo(0, buy_num, pay_how, pay_sum, null, 0);
+		PayVo pvo = new PayVo(0, buy_num, pay_how, pay_sum, null, 1);
 		PayDao pdao = PayDao.getInstance();
 		int n1 = pdao.insert(pvo);
 		if(n1 <= 0) {
